Cast to int

+++ homedisplay/control_milight/management/commands/run_timed.py
@@ -20,8 +20,8 @@ class Command(BaseCommand):
             percent_done = item.percent_done(time)
             if item.action == "evening":
                 print "Setting evening brightness to", ((1-percent_done)*100)
-                led.set_brightness((1-percent_done)*100)
+                led.set_brightness(int((1-percent_done)*100))
             elif item.action == "morning":
                 print "Setting morning brightness to", ((percent_done)*100)
-                led.set_brightness((percent_done)*100)
+                led.set_brightness(int((percent_done)*100))
 #            update_lightstate(transition.group.group_id, transition.to_brightness, transition.to_color)
